backend/requirements.txt
# Flask and Web Server
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0

# Background Jobs
celery==5.4.0
redis==5.0.7

# Subtitle Parsing
pysrt==1.1.2

# AI, ML, and Data Processing
scikit-learn==1.5.0
google-generativeai==0.7.1
numpy==1.26.4
faiss-cpu==1.8.0
tensorflow==2.16.1
tensorflow-hub==0.16.1
nltk==3.8.1